+++ src/Websites/Websites.Test/ScenarioTests/WebsitesTestRunner.cs
@@ -52,6 +52,7 @@ namespace Microsoft.Azure.Commands.Websites.Test.ScenarioTests
                         {"Microsoft.Resources", null},
                         {"Microsoft.Features", null},
                         {"Microsoft.Authorization", null},
+                        {"Microsoft.Web", null},
                         {"Microsoft.Websites", null},
                         {"Microsoft.Network", null},
                         {"Microsoft.KeyVault", null},