+++ src/Websites/Websites/Utilities/NetworkClient.cs
@@ -17,7 +17,6 @@ using Microsoft.Azure.Commands.Common.Authentication.Abstractions;
 using Microsoft.Azure.Management.Internal.Resources.Utilities.Models;
 using Microsoft.Azure.PowerShell.Cmdlets.Websites.Helper.Network;
 using Microsoft.Azure.PowerShell.Cmdlets.Websites.Helper.Network.Models;
-using Microsoft.Azure.Management.WebSites;
 using System;
 using System.Collections.Generic;
 using System.Linq;